requirements.txt
easyocr==1.7.1
matplotlib==3.8.2
numpy==1.26.3
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
Pillow==10.1.0
Pillow==10.2.0
streamlit==1.29.0
ultralytics==8.0.231
